server/ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = ""
authors = [
    {name = "abhay-byte"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "chromadb (>=1.1.1,<2.0.0)",
    "langchain-text-splitters (>=0.3.11,<0.4.0)",
    "vllm (>=0.11.0,<0.12.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "kagglehub (>=0.3.13,<0.4.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)"
]

[tool.poetry]
packages = [{include = "ai_service", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
